add edit form, switch between edit and save
