+++ src/App.js
@@ -1,16 +1,16 @@
 import React from "react";
 import { Container } from "react-bootstrap";
 import "./App.css";
-//import Footer from "./components/Footer/Footer";
+import Footer from "./components/Footer/Footer";
 import "bootstrap/dist/css/bootstrap.min.css";
-import UserProfile from "./components/UserProfile/UserProfile";
+import UserProfile from "./containers/UserProfile/UserProfile";
 
 function App() {
   return (
     <>
       <Container>
-        {/*<Footer />*/}
         <UserProfile />
+        <Footer />
       </Container>
     </>
   );
